requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
dnspython==2.6.1
fastapi==0.115.0
frozenlist==1.4.1
h11==0.14.0
idna==3.10
magic-filter==1.0.12
motor==3.6.0
multidict==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.1
python-dotenv==1.0.1
setuptools==75.1.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.30.6
wheel==0.38.4
yarl==1.12.1

openai~=1.50.2